Add notifications

+++ app/src/main/java/denis/beck/dutyreminder_2/remindManager/RemindReceiver.kt
@@ -4,11 +4,10 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import denis.beck.dutyreminder_2.DutyReminderApp
+import denis.beck.dutyreminder_2.notification.RemindNotificationManager
 import denis.beck.dutyreminder_2.utils.goAsync
-import denis.beck.dutyreminder_2.utils.showToast
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 import timber.log.Timber
+import java.lang.IllegalStateException
 
 class RemindReceiver : BroadcastReceiver() {
 
@@ -19,13 +18,11 @@ class RemindReceiver : BroadcastReceiver() {
     }
 
     override fun onReceive(context: Context?, intent: Intent?) = goAsync {
-        if (intent?.action == REMIND_ACTION) {
+        if (intent?.action == REMIND_ACTION && context != null) {
             val id = intent.getLongExtra(REMIND_ID_EXTRA, -1)
-            val message = intent.getStringExtra(REMIND_MESSAGE_EXTRA)
-            withContext(Dispatchers.Main) {
-                context?.showToast("Remind Action Received with extra: $message")
-            }
-            val remindDao = (context?.applicationContext as DutyReminderApp).remindDatabase.reminderDao()
+            val message = intent.getStringExtra(REMIND_MESSAGE_EXTRA) ?: throw IllegalStateException("message is null")
+            RemindNotificationManager(context).showNotification(id, message)
+            val remindDao = (context.applicationContext as DutyReminderApp).remindDatabase.reminderDao()
             Timber.d("deleted id: $id prpr")
             remindDao.delete(id)
         }
